+++ Cargo.toml
@@ -9,13 +9,17 @@ anyhow = "1.0.55"
 async_zip = "0.0.6"
 bincode = "1.3.3"
 chrono = "0.4.19"
+clap = { version = "3.2.20", features = ["derive", "env"] }
+clap-verbosity-flag = "1.0.1"
 cloud-storage = "0.11.0"
 counter = "0.5.3"
 derive_builder = "0.10.2"
 dotenv = "0.15.0"
+fern = "0.6.1"
 futures = "0.3.24"
 glob = "0.3.0"
 humantime = "2.1.0"
+log = "0.4.17"
 maud = "0.23.0"
 moka = { version = "0.7.2", features = ["future"] }
 mp3-metadata = "0.3.3"
@@ -23,15 +27,16 @@ mp4ameta = "0.11.0"
 nom = "7.1.0"
 notify = { version = "5.0.0", default-features = false, features = ["macos_kqueue"] }
 notify-debouncer-mini = "0.2.1"
-once_cell = "1.10.0"
+once_cell = "1.14.0"
 openssl-probe = "0.1.5"
+parking_lot = "0.12.1"
 prettytable-rs = "0.8.0"
 rand = "0.8.5"
 reqwest = "0.11.9"
 serde = { version = "1.0.136", features = ["derive"] }
 serde_json = "1.0.79"
-serenity = { version = "0.10.10", features = ["cache", "framework", "standard_framework", "voice"] }
-songbird = "0.2.2"
+serenity = { version = "0.11.5", features = ["cache", "framework", "standard_framework", "voice"] }
+songbird = "0.3.0"
 strsim = "0.10.0"
 structopt = "0.3.26"
 systemstat = "0.1.10"
