+++ rustfmt.toml
@@ -1,0 +1,3 @@
+reorder_impl_items = true
+imports_granularity = "Crate"
+group_imports = "StdExternalCrate"
